expand handleAddCollateral; convert fields from BigInt to BigDecimal

+++ src/utils/common.ts
@@ -1,9 +1,9 @@
-import { BigInt, Bytes } from "@graphprotocol/graph-ts"
+import { BigDecimal, BigInt, Bytes } from "@graphprotocol/graph-ts"
 import { Bucket, Lend } from "../../generated/schema"
 
 import { ZERO_BI } from "./constants"
 
-export function lpbValueInQuote(bucket: Bucket, lend: Lend): BigInt {
+export function lpbValueInQuote(bucket: Bucket, lend: Lend): BigDecimal {
     // TODO: need to convert from RAY to WAD
     let quoteTokenAmount = lend.lpb.times(bucket.exchangeRate)
 
@@ -17,6 +17,6 @@ export function encumberance(debt: BigInt, price: BigInt): BigInt {
     return debt.div(price)
 }
 
-export function collateralization(debt: BigInt, encumberedCollateral: BigInt): BigInt {
+export function collateralization(debt: BigDecimal, encumberedCollateral: BigDecimal): BigDecimal {
     return debt.div(encumberedCollateral)
 }
